+++ src/storm/storage/expressions/Expression.h
@@ -60,6 +60,7 @@ namespace storm {
             friend Expression minimum(Expression const& first, Expression const& second);
             friend Expression maximum(Expression const& first, Expression const& second);
 
+
             Expression() = default;
             ~Expression();
             
@@ -439,6 +440,9 @@ namespace storm {
         Expression modulo(Expression const& first, Expression const& second);
         Expression minimum(Expression const& first, Expression const& second);
         Expression maximum(Expression const& first, Expression const& second);
+        Expression atLeastOneOf(std::vector<storm::expressions::Expression> const& expressions);
+        Expression atMostOneOf(std::vector<storm::expressions::Expression> const& expressions);
+        Expression exactlyOneOf(std::vector<storm::expressions::Expression> const& expressions);
         Expression disjunction(std::vector<storm::expressions::Expression> const& expressions);
         Expression conjunction(std::vector<storm::expressions::Expression> const& expressions);
         Expression sum(std::vector<storm::expressions::Expression> const& expressions);